# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Time Archive API
  version: 0.1.0
  description: |
    Development-stage API contract for the Time Archive MVP purchase flow.

    The current API includes development-stage session authentication,
    reservation, checkout, media upload, moderation, and public timeline
    surfaces. It does not implement real payment provider integration or
    payment completion through browser redirects.

    The `buyerId` request field is temporary development-stage behavior. Before
    production, buyer identity must come from authenticated server-side identity.

    Owned range media endpoints temporarily accept `X-User-Id` as a
    development-stage stand-in for authenticated server-side identity.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
  /api/auth/register:
    post:
      tags:
        - Auth
      operationId: registerUser
      summary: Register a local MVP user
      description: |
        Creates a local MVP user account and signs the browser in through a
        server-side session cookie.

        This endpoint is intentionally minimal and does not perform email
        verification or password reset flows.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered and signed in.
          headers:
            Set-Cookie:
              schema:
                type: string
              description: HttpOnly server session cookie.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUserResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "409":
          description: Email is already registered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "500":
          $ref: "#/components/responses/UnexpectedError"
  /api/auth/login:
    post:
      tags:
        - Auth
      operationId: loginUser
      summary: Log in with email and password
      description: |
        Authenticates a local MVP user and signs the browser in through a
        server-side session cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: User authenticated and signed in.
          headers:
            Set-Cookie:
              schema:
                type: string
              description: HttpOnly server session cookie.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUserResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/UnexpectedError"
  /api/auth/logout:
    post:
      tags:
        - Auth
      operationId: logoutUser
      summary: Log out current session
      description: Invalidates the current server-side session if one exists.
      security:
        - SessionCookie: []
      responses:
        "204":
          description: Session invalidated.
        "500":
          $ref: "#/components/responses/UnexpectedError"
  /api/me:
    get:
      tags:
        - Auth
      operationId: getCurrentUser
      summary: Get current user
      description: Returns the current authenticated user derived from the server-side session.
      security:
        - SessionCookie: []
      responses:
        "200":
          description: Current authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/UnexpectedError"
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
...
  responses:
    InvalidRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
...
    Unauthorized:
      description: Authentication is required or credentials are invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    UnexpectedError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
  schemas:
    RegisterRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
        - displayName
      properties:
        email:
          type: string
          format: email
          minLength: 1
          examples:
            - user@example.com
        password:
          type: string
          minLength: 8
          maxLength: 128
          writeOnly: true
        displayName:
          type: string
          minLength: 1
          maxLength: 80
          examples:
            - User
    LoginRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          minLength: 1
          examples:
            - user@example.com
        password:
          type: string
          minLength: 8
          maxLength: 128
          writeOnly: true
    CurrentUserResponse:
      type: object
      additionalProperties: false
      required:
        - userId
        - email
        - displayName
        - createdAt
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        createdAt:
          type: string
          format: date-time
...
    ApiErrorResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - details
      properties:
        code:
          $ref: "#/components/schemas/ApiErrorCode"
        message:
          type: string
          minLength: 1
        details:
          type: array
          items:
            $ref: "#/components/schemas/ApiErrorDetail"
    ApiErrorCode:
      type: string
      enum:
        - INVALID_REQUEST
        - RESOURCE_NOT_FOUND
        - TIME_RANGE_ALREADY_OWNED
        - TIME_RANGE_ALREADY_RESERVED
        - RESERVATION_EXPIRED
        - RESERVATION_NOT_PAYABLE
        - PAYMENT_EVENT_ALREADY_PROCESSING
        - OWNERSHIP_ACCESS_DENIED
        - OWNERSHIP_NOT_ACTIVE
        - MEDIA_UPLOAD_REQUEST_EXPIRED
        - MEDIA_UPLOAD_REQUEST_NOT_COMPLETABLE
        - MEDIA_UPLOAD_OBJECT_NOT_FOUND
        - MEDIA_UPLOAD_OBJECT_MISMATCH
        - MEDIA_ASSET_NOT_APPROVABLE
        - MEDIA_ASSET_NOT_REJECTABLE
        - MEDIA_ASSET_NOT_HIDEABLE
        - UNEXPECTED_ERROR
    ApiErrorDetail:
      type: object
      additionalProperties: false
      required:
        - field
        - message
      properties:
        field:
          type: string
          description: Field path related to the error. Use `null` for non-field errors.
          nullable: true
          examples:
            - startSecond
        message:
          type: string
          minLength: 1